requirements.txt
faust-streaming==0.10.16
pymongo==4.5.0
python-dotenv==1.0.0
redis==5.0.1
kafka-python==2.0.2
mysql-connector-python==8.1.0